+++ kamehouse-shell/bin/kamehouse/docker/docker-ci-integration-tests-trigger.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Config in ci to run as: sudo -u [USERNAME] /home/[USERNAME]/programs/kamehouse-shell/bin/kamehouse/docker/docker-ci-integration-tests-trigger.sh
-# Add to sudoers:
-# jenkins ALL=(ALL) NOPASSWD: /home/[USERNAME]/programs/kamehouse-shell/bin/kamehouse/docker/docker-ci-integration-tests-trigger.sh
-
 # Import common functions
 source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
 if [ "$?" != "0" ]; then